+++ packages/elevenlabs-client/src/with-elevenlabs.ts
@@ -1,11 +1,13 @@
 import { TEnforceFeatureConstraint, TFeatureDefinition } from '@blgc/types/features';
-import { mapOk, Ok, unwrapOrNull } from '@blgc/utils';
 import {
 	Err,
 	FetchError,
 	isFetchClientWithFeatures,
+	mapOk,
+	Ok,
 	TFetchClient,
-	TOpenApiFeature
+	TOpenApiFeature,
+	unwrapOrNull
 } from 'feature-fetch';
 import { type paths } from './gen/v1';
 import { isVoiceId } from './helper';