+++ packages/feature-fetch/src/create-fetch-client.ts
@@ -1,4 +1,4 @@
-import { Err, Ok } from '@blgc/utils';
+import { Err, Ok } from 'tuple-result';
 import { FetchError } from './exceptions';
 import {
 	buildUrl,