+++ packages/feature-fetch/src/index.ts
@@ -4,4 +4,5 @@ export * from './features';
 export * from './helper';
 export * from './types';
 
-export { Err, Ok, type TErrResult, type TOkResult, type TResult } from '@blgc/utils';
+// Re-export tuple-result for convenience
+export * from 'tuple-result';