+++ packages/google-webfonts-client/src/with-google-webfonts.ts
@@ -1,10 +1,10 @@
 import { TEnforceFeatureConstraint, TFeatureDefinition } from '@blgc/types/features';
-import { mapOk } from '@blgc/utils';
 import {
 	createApiFetchClient,
 	Err,
 	isFetchClientWithFeatures,
 	isStatusCode,
+	mapOk,
 	Ok,
 	TOpenApiFeature,
 	type TFetchClient