+++ packages/utils/src/index.ts
@@ -35,7 +35,6 @@ export * from './not-empty';
 export * from './pick-properties';
 export * from './rad-to-deg';
 export * from './random-hex';
-export * from './result';
 export * from './rgb-to-hex';
 export * from './rgb-to-rgba';
 export * from './rgba-to-rgb';