+++ packages/ecsify/CHANGELOG.md
@@ -1,5 +1,14 @@
 # ecsify
 
+## 0.0.13
+
+### Patch Changes
+
+- updated deps
+
+- Updated dependencies []:
+  - @blgc/utils@0.0.59
+
 ## 0.0.12
 
 ### Patch Changes