+++ packages/head-metadata/CHANGELOG.md
@@ -1,5 +1,14 @@
 # head-metadata
 
+## 0.0.9
+
+### Patch Changes
+
+- updated deps
+
+- Updated dependencies []:
+  - xml-tokenizer@0.0.42
+
 ## 0.0.8
 
 ### Patch Changes