+++ packages/openapi-ts-router/CHANGELOG.md
@@ -1,5 +1,15 @@
 # openapi-ts-router
 
+## 0.3.4
+
+### Patch Changes
+
+- updated deps
+
+- Updated dependencies []:
+  - validation-adapter@0.0.25
+  - @blgc/types@0.0.19
+
 ## 0.3.3
 
 ### Patch Changes