+++ packages/rollup-presets/CHANGELOG.md
@@ -1,5 +1,11 @@
 # rollup-presets
 
+## 0.0.23
+
+### Patch Changes
+
+- updated deps
+
 ## 0.0.22
 
 ### Patch Changes