+++ packages/webito/CHANGELOG.md
@@ -1,5 +1,14 @@
 # webito
 
+## 0.0.4
+
+### Patch Changes
+
+- updated deps
+
+- Updated dependencies []:
+  - ecsify@0.0.13
+
 ## 0.0.3
 
 ### Patch Changes